somewhat safer typecasting in case sizeof(long) != sizeof(void *) (is there even such platforms?)

+++ lib/hostip.c
@@ -1012,11 +1012,11 @@ static struct hostent* pack_hostent(char** buf, struct hostent* orig)
 
   /* now, shrink the allocated buffer to the size we actually need, which
      most often is only a fraction of the original alloc */
-  newbuf=(char *)realloc(*buf, (long)bufptr-(long)(*buf));
+  newbuf=(char *)realloc(*buf, (long)(bufptr-*buf));
 
   /* if the alloc moved, we need to adjust things again */
   if(newbuf != *buf)
-    hostcache_fixoffset((struct hostent*)newbuf, (long)newbuf-(long)*buf);
+    hostcache_fixoffset((struct hostent*)newbuf, (long)(newbuf-*buf));
 
   /* setup the return */
   *buf = newbuf;
